+++ packages/core/src/components/CharacterViewsPreview/index.ts
@@ -1,0 +1,1 @@
+export { default as CharacterViewsPreview } from './CharacterViewsPreview.vue'